requirements.txt
transformers
torch
groq
python-dotenv
streamlit
nest_asyncio